requirements.txt
amqp==5.0.6
arabic-reshaper==2.1.3
asgiref==3.4.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.1.2
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.8
django-celery-results==2.2.0
django-countries==7.2.1
django-jalali==4.3.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
future==0.18.2
html5lib==1.1
importlib-metadata==4.8.1
jdatetime==3.6.4
kombu==5.1.0
Pillow==8.3.2
prompt-toolkit==3.0.20
psycopg2-binary==2.9.1
PyJWT==1.7.1
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2021.3
reportlab==3.6.1
six==1.16.0
sqlparse==0.4.2
typing-extensions==3.10.0.2
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
xhtml2pdf==0.2.5
zipp==3.6.0
